requirements.txt
Flask==2.1.2
flask-restx==0.5.1
flask-mongoengine==1.0.0
Flask-JWT-Extended==4.4.1
marshmallow==3.16.0
webargs==8.1.0
opencv-python==4.6.0.66
cmake==3.22.5
face-recognition==1.3.0
sklearn
argon2-cffi==21.3.0